requirements.txt
# Core AI and API
openai>=1.53.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
gunicorn==21.2.0

# Optional: Flask (if you still use it elsewhere)
flask>=3.0.0

# CORS support for frontend/backend communication
fastapi[all]>=0.110.0

# Optional: For streaming responses with Server-Sent Events
sse-starlette>=1.6.5

# Optional: For async OpenAI calls
httpx>=0.27.0